Docs : move component
